Classe Relatorio
Uma classe relatório foi atribuída com os métodos (relatório de estoque baixo, Relatório Geral e Relatório de Venda), para não sobre carregar a classe controle de estoque.
